requirements.txt
discord.py
replit.py
requests.py
